requirements.txt
snscrape
google-generativeai
